requirements.txt
python-telegram-bot==20.7
spotipy==2.23.0
requests==2.31.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
torch==2.0.1
pandas==2.0.3
matplotlib==3.7.2
scikit-learn==1.3.0
faiss-cpu==1.7.4
numpy==1.24.3
textstat==0.7.3
joblib==1.3.2